# badges-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Badges API
  description: An API to demonstrate Swagger and Spring Boot
security:
  #  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /badges:
    post:
      description: create a badge
      operationId: createBadge
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
    get:
      description: get the list of all badges
      operationId: getBadges
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
  /badges/{id}:
    get:
      description: get one badge, by id
      operationId: getBadge
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
  /badges/user/{username}:
    get:
      description: get badges for one user
      operationId: getUserBadges
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
  /applications:
    post:
      description: register your app to the gamification server
      operationId: registerApp
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Application'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ApiKey'
  /events:
    post:
      description: send an event to the gamification server
      operationId: sendEvent
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: proceed
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
  /rules:
    post:
      description: add a new rule for the application
      operationId: createRule
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Rule'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Rule'
    get:
      description: get the list of all rules
      operationId: getRules
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
  /rules/{id}:
    get:
      description: get one rule, by id
      operationId: getRule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Rule'
  /eventCounters:
    post:
      description: add a new event counter for the application
      operationId: createEventCounter
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/EventCounter'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/EventCounter'
    get:
      description: get the list of all event counters
      operationId: getEventCounters
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventCounter'
  /eventCounters/{id}:
      get:
        description: get one event counter, by id
        operationId: getEventCounter
        parameters:
          - in: path
            name: id
            schema:
              type: integer
            required: true
        responses:
          '200':
            description: success
            content:
              'application/json':
                schema:
                  $ref: '#/components/schemas/EventCounter'
  /top/{n}:
    get:
      description: get the top n the users ordered by their total experience value
      operationId: getTopN
      parameters:
        - in: path
          name: n
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'


components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    Badge:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        experienceValue:
          type: integer
    Application:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ApiKey:
      type: object
      properties:
        uuid:
          type: string
    Event:
      type: object
      properties:
        username:
          type: string
        eventCounterId:
          type: integer
          format: int64
    Rule:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventCounterId:
          type: integer
          format: int64
        threshold:
          type: integer
        badgeId:
          type: integer
    EventCounter:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        initialValue:
          type: integer
    User:
      type: object
      properties:
        username:
          type: string
        experienceValue:
          type: integer
          format: int64